+++ grid2op/tests/test_multi_steps_forecasts.py
@@ -31,13 +31,17 @@ class MultiStepsForcaTester(unittest.TestCase):
         self.env.set_id(0)
     
     def aux_test_for_consistent(self, obs):
-        tmp_o_1, *_ = obs.simulate(self.env.action_space(), time_step=1)
+        tmp_o_1, *_ = obs.simulate(self.env.action_space(),
+                                   time_step=1, chain_independant=True)
         assert (obs.load_p + 1. == tmp_o_1.load_p).all()  # that's how I generated the forecast for this "env"
-        tmp_o_2, *_ = obs.simulate(self.env.action_space(), time_step=2)
+        tmp_o_2, *_ = obs.simulate(self.env.action_space(),
+                                   time_step=2, chain_independant=True)
         assert (obs.load_p + 2. == tmp_o_2.load_p).all()  # that's how I generated the forecast for this "env"
-        tmp_o_3, *_ = obs.simulate(self.env.action_space(), time_step=3)
+        tmp_o_3, *_ = obs.simulate(self.env.action_space(),
+                                   time_step=3, chain_independant=True)
         assert (obs.load_p + 3. == tmp_o_3.load_p).all()
-        tmp_o_12, *_ = obs.simulate(self.env.action_space(), time_step=12)
+        tmp_o_12, *_ = obs.simulate(self.env.action_space(),
+                                    time_step=12, chain_independant=True)
         assert (obs.load_p + 12. == tmp_o_12.load_p).all()
         
     def test_can_do(self):
@@ -46,7 +50,9 @@ class MultiStepsForcaTester(unittest.TestCase):
         
         # should raise because there is no "13 steps ahead forecasts"
         with self.assertRaises(NoForecastAvailable):
-            obs.simulate(self.env.action_space(), time_step=13)
+            obs.simulate(self.env.action_space(),
+                         time_step=13,
+                         chain_independant=True)
         
         # check it's still consistent
         obs, *_ = self.env.step(self.env.action_space())
@@ -124,5 +130,89 @@ class MultiStepsForcaTester(unittest.TestCase):
         self.aux_test_for_consistent(obs)
 
 
+class ChainSimulateTester(unittest.TestCase):
+    def setUp(self) -> None:
+        with warnings.catch_warnings():
+            warnings.filterwarnings("ignore")
+            # this needs to be tested with pandapower backend
+            self.env = grid2op.make(os.path.join(PATH_DATA_TEST, "5bus_example_forecasts"), test=True)
+        self.env.seed(0)
+        self.env.set_id(0)
+    
+    def aux_test_for_consistent_independant(self, obs, tmp_o, h):
+        assert (obs.load_p + 1. * h == tmp_o.load_p).all()
+        tmp_o_1, *_ = obs.simulate(self.env.action_space(),
+                                   time_step=h,
+                                   chain_independant=True)
+        assert (obs.load_p + 1. * h == tmp_o_1.load_p).all()
+         
+    def test_can_chain_independant(self):
+        obs = self.env.reset()
+        tmp_o_1, *_ = obs.simulate(self.env.action_space(),
+                                   time_step=1,
+                                   chain_independant=True)
+        self.aux_test_for_consistent(obs, tmp_o_1, 1)
+        
+        tmp_o_2, *_ = tmp_o_1.simulate(self.env.action_space(),
+                                       time_step=1,
+                                       chain_independant=True)
+        self.aux_test_for_consistent(obs, tmp_o_2, 2)
+        _ = tmp_o_1.simulate(self.env.action_space(),
+                             time_step=11,
+                             chain_independant=True)
+        with self.assertRaises(NoForecastAvailable):
+            # not available
+            tmp_o_2, *_ = tmp_o_1.simulate(self.env.action_space(),
+                                           time_step=12,
+                                           chain_independant=True)
+        
+        tmp_o_3, *_ = tmp_o_2.simulate(self.env.action_space(),
+                                       time_step=1,
+                                       chain_independant=True)
+        self.aux_test_for_consistent(obs, tmp_o_3, 3)
+        _ = tmp_o_2.simulate(self.env.action_space(),
+                             time_step=10,
+                             chain_independant=True)
+        with self.assertRaises(NoForecastAvailable):
+            # not available
+            tmp_o_2, *_ = tmp_o_2.simulate(self.env.action_space(),
+                                           time_step=11,
+                                           chain_independant=True)
+    
+    def test_can_chain_dependant(self):
+        obs = self.env.reset()
+        dn = self.env.action_space()
+        # if I do nothing it's like it's independant
+        tmp_o_1, *_ = obs.simulate(dn, time_step=1)
+        self.aux_test_for_consistent_independant(obs, tmp_o_1, 1)
+        
+        # check that it's not independant
+        act = self.env.action_space({"set_bus": {"substations_id": [(2, (2, 1, 2, 1))]}})
+        tmp_o_1_1, *_ = obs.simulate(act, time_step=1)
+        assert (tmp_o_1_1.topo_vect[[9, 11]] == [2, 2]).all()
+        tmp_o_2_1, *_ = tmp_o_1_1.simulate(dn, time_step=1)
+        assert (tmp_o_2_1.topo_vect[[9, 11]] == [2, 2]).all()
+        # check that the original simulate is not "broken"
+        tmp_o_1_base, *_ = obs.simulate(dn, time_step=1)
+        assert (tmp_o_1_base.topo_vect[[9, 11]] == [1, 1]).all()
+        
+        # and to be sure, check that it's independant if i put the flag
+        # it's surprising that it works TODO !
+        act = self.env.action_space({"set_bus": {"substations_id": [(2, (2, 1, 2, 1))]}})
+        tmp_o_1_2, *_ = obs.simulate(act, time_step=1, chain_independant=True)
+        assert (tmp_o_1_2.topo_vect[[9, 11]] == [2, 2]).all()
+        # check that the original simulate is not "broken"
+        tmp_o_1_base, *_ = obs.simulate(dn, time_step=1, chain_independant=True)
+        assert (tmp_o_1_base.topo_vect[[9, 11]] == [1, 1]).all()
+        # check 2nd simulate is indpendant of first one
+        tmp_o_2_2, *_ = tmp_o_1_2.simulate(dn, time_step=1, chain_independant=True)
+        assert (tmp_o_2_2.topo_vect[[9, 11]] == [2, 2]).all()
+        # check that the original simulate is not "broken"
+        tmp_o_1_base, *_ = obs.simulate(dn, time_step=1)
+        assert (tmp_o_1_base.topo_vect[[9, 11]] == [1, 1]).all()
+        
+        
+# TODO check cooldowns
+# TODO check "thermal limit when soft overflow"
 if __name__ == "__main__":
     unittest.main()